refactor version 1 many todo
