+++ demo-shell/src/app/components/cloud/cloud-viewer.component.html
@@ -1,4 +1,3 @@
-<adf-viewer
-    [overlayMode]="true"
+<adf-alfresco-viewer
     [nodeId]="nodeId">
-</adf-viewer>
+</adf-alfresco-viewer>
