+++ lib/core/src/lib/services/url.service.ts
@@ -1,0 +1,42 @@
+/*!
+ * @license
+ * Copyright 2019 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Injectable } from '@angular/core';
+import { DomSanitizer } from '@angular/platform-browser';
+
+@Injectable({
+    providedIn: 'root'
+})
+export class UrlService {
+
+    constructor(private sanitizer: DomSanitizer) {
+    }
+
+    /**
+     * Creates a trusted object URL from the Blob.
+     * WARNING: calling this method with untrusted user data exposes your application to XSS security risks!
+     *
+     * @param  blob Data to wrap into object URL
+     * @returns URL string
+     */
+    createTrustedUrl(blob: Blob): string {
+        const url = window.URL.createObjectURL(blob);
+        return this.sanitizer.bypassSecurityTrustUrl(url) as string;
+    }
+
+
+}
