+++ lib/core/src/lib/viewer/components/media-player.component.ts
@@ -18,7 +18,6 @@
 import { Component, Input, OnChanges, SimpleChanges, ViewEncapsulation, Output, EventEmitter } from '@angular/core';
 import { ContentService } from '../../services/content.service';
 import { Track } from '../models/viewer.model';
-import { ViewUtilService } from '../services/view-util.service';
 
 @Component({
     selector: 'adf-media-player',
@@ -39,10 +38,10 @@ export class MediaPlayerComponent implements OnChanges {
     mimeType: string;
 
     @Input()
-    nameFile: string;
-
-    @Input()
-    nodeId: string;
+    fileName: string;
+    //
+    // @Input()
+    // nodeId: string;
 
     @Input()
     tracks: Track[] = [];
@@ -50,21 +49,21 @@ export class MediaPlayerComponent implements OnChanges {
     @Output()
     error = new EventEmitter<any>();
 
-    constructor(private contentService: ContentService, private viewUtils: ViewUtilService) {
+    constructor(private contentService: ContentService) {
     }
 
     ngOnChanges(changes: SimpleChanges) {
         const blobFile = changes['blobFile'];
-        const nodeId = changes['nodeId'];
+       // const nodeId = changes['nodeId'];
 
         if (blobFile && blobFile.currentValue) {
             this.urlFile = this.contentService.createTrustedUrl(this.blobFile);
             return;
         }
 
-        if (nodeId && nodeId.currentValue) {
-            this.viewUtils.generateMediaTracks(this.nodeId).then((tracks) => this.tracks = tracks);
-        }
+        // if (nodeId && nodeId.currentValue) {
+        //     this.viewUtils.generateMediaTracksRendition(this.nodeId).then((tracks) => this.tracks = tracks);
+        // }
 
         if (!this.urlFile && !this.blobFile) {
             throw new Error('Attribute urlFile or blobFile is required');
