+++ lib/core/src/lib/viewer/services/view-util.service.ts
@@ -16,12 +16,9 @@
  */
 
 import { Injectable } from '@angular/core';
-import { ContentApi, RenditionEntry, RenditionPaging, RenditionsApi, VersionsApi } from '@alfresco/js-api';
-import { AlfrescoApiService } from '../../services/alfresco-api.service';
 import { LogService } from '../../common/services/log.service';
-import { Subject } from 'rxjs';
-import { Track } from '../models/viewer.model';
-import { TranslationService } from '../../translation/translation.service';
+import { RenditionViewerService } from "./rendition-viewer.service";
+import { AppExtensionService, ViewerExtensionRef } from '@alfresco/adf-extensions';
 
 @Injectable({
     providedIn: 'root'
@@ -29,70 +26,69 @@ import { TranslationService } from '../../translation/translation.service';
 export class ViewUtilService {
     static TARGET = '_new';
 
-    /**
-     * Content groups based on categorization of files that can be viewed in the web browser. This
-     * implementation or grouping is tied to the definition the ng component: ViewerComponent
-     */
-    static ContentGroup = {
-        IMAGE: 'image',
-        MEDIA: 'media',
-        PDF: 'pdf',
-        TEXT: 'text'
+    // Extensions that are supported by the Viewer without conversion
+    private extensions = {
+        image: ['png', 'jpg', 'jpeg', 'gif', 'bpm', 'svg'],
+        media: ['wav', 'mp4', 'mp3', 'webm', 'ogg'],
+        text: ['txt', 'xml', 'html', 'json', 'ts', 'css', 'md'],
+        pdf: ['pdf']
     };
 
-    /**
-     * The name of the rendition with the media subtitles in the supported format
-     */
-    static SUBTITLES_RENDITION_NAME = 'webvtt';
-
-    /**
-     * Based on ViewerComponent Implementation, this value is used to determine how many times we try
-     * to get the rendition of a file for preview, or printing.
-     */
-    maxRetries = 5;
-
-    /**
-     * Mime-type grouping based on the ViewerComponent.
-     */
+    // Mime types that are supported by the Viewer without conversion
     private mimeTypes = {
         text: ['text/plain', 'text/csv', 'text/xml', 'text/html', 'application/x-javascript'],
         pdf: ['application/pdf'],
         image: ['image/png', 'image/jpeg', 'image/gif', 'image/bmp', 'image/svg+xml'],
-        media: ['video/mp4', 'video/webm', 'video/ogg', 'audio/mpeg', 'audio/ogg', 'audio/wav']
+        media: ['video/mp4', 'video/webm', 'video/ogg', 'audio/mpeg', 'audio/mp3', 'audio/ogg', 'audio/wav']
     };
 
     /**
-     * Timeout used for setInterval.
+     * Returns a list of the active Viewer content extensions.
      */
-    TRY_TIMEOUT: number = 10000;
+    get viewerExtensions(): ViewerExtensionRef[] {
+        return this.extensionService.getViewerExtensions();
+    }
 
     /**
-     * Subscribers needed for ViewerComponent to update the viewerType and urlFileContent.
+     * Provides a list of file extensions supported by external plugins.
      */
-    viewerTypeChange: Subject<string> = new Subject<string>();
-    urlFileContentChange: Subject<string> = new Subject<string>();
-
-    private _renditionsApi: RenditionsApi;
-    get renditionsApi(): RenditionsApi {
-        this._renditionsApi = this._renditionsApi ?? new RenditionsApi(this.apiService.getInstance());
-        return this._renditionsApi;
+    get externalExtensions(): string[] {
+        return this.viewerExtensions.map(ext => ext.fileExtension);
     }
 
-    private _contentApi: ContentApi;
-    get contentApi(): ContentApi {
-        this._contentApi = this._contentApi ?? new ContentApi(this.apiService.getInstance());
-        return this._contentApi;
+    constructor(private renditionViewerService: RenditionViewerService,
+                private extensionService: AppExtensionService,
+                private logService: LogService) {
     }
 
-    private _versionsApi: VersionsApi;
-    get versionsApi(): VersionsApi {
-        this._versionsApi = this._versionsApi ?? new VersionsApi(this.apiService.getInstance());
-        return this._versionsApi;
+    /**
+     * get File name from url
+     *
+     * @param  url - url file
+     */
+    getFilenameFromUrl(url: string): string {
+        const anchor = url.indexOf('#');
+        const query = url.indexOf('?');
+        const end = Math.min(
+            anchor > 0 ? anchor : url.length,
+            query > 0 ? query : url.length);
+        return url.substring(url.lastIndexOf('/', end) + 1, end);
     }
 
-    constructor(private apiService: AlfrescoApiService,
-                private logService: LogService,
-                private translateService: TranslationService) {
+    /**
+     * Get file extension from the string.
+     * Supports the URL formats like:
+     * http://localhost/test.jpg?cache=1000
+     * http://localhost/test.jpg#cache=1000
+     *
+     * @param fileName - file name
+     */
+    getFileExtension(fileName: string): string {
+        if (fileName) {
+            const match = fileName.match(/\.([^\./\?\#]+)($|\?|\#)/);
+            return match ? match[1] : null;
+        }
+        return null;
     }
 
     /**
@@ -104,7 +100,7 @@ export class ViewUtilService {
         const pwa = window.open(url, ViewUtilService.TARGET);
         if (pwa) {
             // Because of the way chrome focus and close image window vs. pdf preview window
-            if (type === ViewUtilService.ContentGroup.IMAGE) {
+            if (type === RenditionViewerService.ContentGroup.IMAGE) {
                 pwa.onfocus = () => {
                     setTimeout(() => {
                         pwa.close();
@@ -127,14 +123,14 @@ export class ViewUtilService {
      */
     printFileGeneric(objectId: string, mimeType: string): void {
         const nodeId = objectId;
-        const type: string = this.getViewerTypeByMimeType(mimeType);
+        const type: string = this.renditionViewerService.getViewerTypeByMimeType(mimeType);
 
-        this.getRendition(nodeId, ViewUtilService.ContentGroup.PDF)
+        this.renditionViewerService.getRendition(nodeId, RenditionViewerService.ContentGroup.PDF)
             .then((value) => {
-                const url: string = this.getRenditionUrl(nodeId, type, (!!value));
-                const printType = (type === ViewUtilService.ContentGroup.PDF
-                    || type === ViewUtilService.ContentGroup.TEXT)
-                    ? ViewUtilService.ContentGroup.PDF : type;
+                const url: string = this.renditionViewerService.getRenditionUrl(nodeId, type, (!!value));
+                const printType = (type === RenditionViewerService.ContentGroup.PDF
+                    || type === RenditionViewerService.ContentGroup.TEXT)
+                    ? RenditionViewerService.ContentGroup.PDF : type;
                 this.printFile(url, printType);
             })
             .catch((err) => {
@@ -143,31 +139,18 @@ export class ViewUtilService {
             });
     }
 
-    getRenditionUrl(nodeId: string, type: string, renditionExists: boolean): string {
-        return (renditionExists && type !== ViewUtilService.ContentGroup.IMAGE) ?
-            this.contentApi.getRenditionUrl(nodeId, ViewUtilService.ContentGroup.PDF) :
-            this.contentApi.getContentUrl(nodeId, false);
-    }
 
-    private async waitRendition(nodeId: string, renditionId: string, retries: number): Promise<RenditionEntry> {
-        const rendition = await this.renditionsApi.getRendition(nodeId, renditionId);
+     getViewerType(extension: string, mimeType: string): string {
+        let viewerType = this.getViewerTypeByExtension(extension);
 
-        if (this.maxRetries < retries) {
-            const status = rendition.entry.status.toString();
-
-            if (status === 'CREATED') {
-                return rendition;
-            } else {
-                retries += 1;
-                await this.wait(1000);
-                return this.waitRendition(nodeId, renditionId, retries);
-            }
+        if (viewerType === 'unknown') {
+            viewerType = this.getViewerTypeByMimeType(mimeType);
         }
 
-        return Promise.resolve(null);
+        return viewerType;
     }
 
-    getViewerTypeByMimeType(mimeType: string): string {
+    getViewerTypeByMimeType(mimeType: string) {
         if (mimeType) {
             mimeType = mimeType.toLowerCase();
 
@@ -181,164 +164,53 @@ export class ViewUtilService {
         return 'unknown';
     }
 
-    wait(ms: number): Promise<any> {
-        return new Promise((resolve) => setTimeout(resolve, ms));
-    }
-
-    async getRendition(nodeId: string, renditionId: string): Promise<RenditionEntry> {
-        const renditionPaging: RenditionPaging = await this.renditionsApi.listRenditions(nodeId);
-        let rendition: RenditionEntry = renditionPaging.list.entries.find((renditionEntry: RenditionEntry) => renditionEntry.entry.id.toLowerCase() === renditionId);
-
-        if (rendition) {
-            const status = rendition.entry.status.toString();
-
-            if (status === 'NOT_CREATED') {
-                try {
-                    await this.renditionsApi.createRendition(nodeId, { id: renditionId });
-                    rendition = await this.waitRendition(nodeId, renditionId, 0);
-                } catch (err) {
-                    this.logService.error(err);
-                }
-            }
-        }
-        return new Promise<RenditionEntry>((resolve) => resolve(rendition));
-    }
-
-    async displayNodeRendition(nodeId: string, versionId?: string) {
-        try {
-            const rendition = versionId ? await this.resolveNodeRendition(nodeId, 'pdf', versionId) :
-                await this.resolveNodeRendition(nodeId, 'pdf');
-            if (rendition) {
-                const renditionId = rendition.entry.id;
-
-                if (renditionId === 'pdf') {
-                    this.viewerTypeChange.next('pdf');
-                } else if (renditionId === 'imgpreview') {
-                    this.viewerTypeChange.next('image');
-                }
-
-                const urlFileContent = versionId ? this.contentApi.getVersionRenditionUrl(nodeId, versionId, renditionId) :
-                    this.contentApi.getRenditionUrl(nodeId, renditionId);
-                this.urlFileContentChange.next(urlFileContent);
-            }
-        } catch (err) {
-            this.logService.error(err);
-        }
-    }
-
-    private async resolveNodeRendition(nodeId: string, renditionId: string, versionId?: string): Promise<RenditionEntry> {
-        renditionId = renditionId.toLowerCase();
-
-        const supportedRendition: RenditionPaging = versionId ? await this.versionsApi.listVersionRenditions(nodeId, versionId) :
-            await this.renditionsApi.listRenditions(nodeId);
-
-        let rendition: RenditionEntry = supportedRendition.list.entries.find((renditionEntry: RenditionEntry) => renditionEntry.entry.id.toLowerCase() === renditionId);
-        if (!rendition) {
-            renditionId = 'imgpreview';
-            rendition = supportedRendition.list.entries.find((renditionEntry: RenditionEntry) => renditionEntry.entry.id.toLowerCase() === renditionId);
+    private getViewerTypeByExtension(extension: string): string {
+        if (extension) {
+            extension = extension.toLowerCase();
         }
 
-        if (rendition) {
-            const status: string = rendition.entry.status.toString();
-
-            if (status === 'NOT_CREATED') {
-                try {
-                    if (versionId) {
-                        await this.versionsApi.createVersionRendition(nodeId, versionId, { id: renditionId }).then(() => {
-                            this.viewerTypeChange.next('in_creation');
-                        });
-                    } else {
-                        await this.renditionsApi.createRendition(nodeId, { id: renditionId }).then(() => {
-                            this.viewerTypeChange.next('in_creation');
-                        });
-                    }
-                    try {
-                        rendition = versionId ? await this.waitNodeRendition(nodeId, renditionId, versionId) : await this.waitNodeRendition(nodeId, renditionId);
-                    } catch (e) {
-                        this.viewerTypeChange.next('error_in_creation');
-                        rendition = null;
-                    }
-
-                } catch (err) {
-                    this.logService.error(err);
-                }
-            }
+        if (this.isExternalViewer()) {
+            return 'external';
         }
 
-        return rendition;
-    }
-
-    private async waitNodeRendition(nodeId: string, renditionId: string, versionId?: string): Promise<RenditionEntry> {
-        let currentRetry: number = 0;
-        return new Promise<RenditionEntry>((resolve, reject) => {
-            const intervalId = setInterval(() => {
-                currentRetry++;
-                if (this.maxRetries >= currentRetry) {
-                    if (versionId) {
-                        this.versionsApi.getVersionRendition(nodeId, versionId, renditionId).then((rendition: RenditionEntry) => {
-                            const status: string = rendition.entry.status.toString();
-
-                            if (status === 'CREATED') {
-                                this.handleNodeRendition(nodeId, renditionId, versionId);
-                                clearInterval(intervalId);
-                                return resolve(rendition);
-                            }
-                        }, () => reject());
-                    } else {
-                        this.renditionsApi.getRendition(nodeId, renditionId).then((rendition: RenditionEntry) => {
-                            const status: string = rendition.entry.status.toString();
-
-                            if (status === 'CREATED') {
-                                this.handleNodeRendition(nodeId, renditionId);
-                                clearInterval(intervalId);
-                                return resolve(rendition);
-                            }
-                        }, () => reject());
-                    }
-                } else {
-                    clearInterval(intervalId);
-                    return reject();
-                }
-            }, this.TRY_TIMEOUT);
-        });
-    }
-
-    private async handleNodeRendition(nodeId: string, renditionId: string, versionId?: string) {
-        if (renditionId === 'pdf') {
-            this.viewerTypeChange.next('pdf');
-        } else if (renditionId === 'imgpreview') {
-            this.viewerTypeChange.next('image');
+        if (this.isCustomViewerExtension(extension)) {
+            return 'custom';
         }
 
-        const urlFileContent = versionId ? this.contentApi.getVersionRenditionUrl(nodeId, versionId, renditionId) :
-            this.contentApi.getRenditionUrl(nodeId, renditionId);
-        this.urlFileContentChange.next(urlFileContent);
+        if (this.extensions.image.indexOf(extension) >= 0) {
+            return 'image';
+        }
+
+        if (this.extensions.media.indexOf(extension) >= 0) {
+            return 'media';
+        }
+
+        if (this.extensions.text.indexOf(extension) >= 0) {
+            return 'text';
+        }
+
+        if (this.extensions.pdf.indexOf(extension) >= 0) {
+            return 'pdf';
+        }
+
+        return 'unknown';
     }
 
-    async generateMediaTracks(nodeId: string): Promise<Track[]> {
-        return this.isRenditionAvailable(nodeId, ViewUtilService.SUBTITLES_RENDITION_NAME)
-            .then((value) => {
-                const tracks = [];
-                if (value) {
-                    tracks.push({
-                        kind: 'subtitles',
-                        src: this.contentApi.getRenditionUrl(nodeId, ViewUtilService.SUBTITLES_RENDITION_NAME),
-                        label: this.translateService.instant('ADF_VIEWER.SUBTITLES')
-                    });
-                }
-                return tracks;
-            })
-            .catch((err) => {
-                this.logService.error('Error while retrieving ' + ViewUtilService.SUBTITLES_RENDITION_NAME + ' rendition');
-                this.logService.error(err);
-                return [];
-            });
+
+    private isExternalViewer(): boolean {
+        return !!this.viewerExtensions.find(ext => ext.fileExtension === '*');
     }
 
-    private async isRenditionAvailable(nodeId: string, renditionId: string): Promise<boolean> {
-        const renditionPaging: RenditionPaging = await this.renditionsApi.listRenditions(nodeId);
-        const rendition: RenditionEntry = renditionPaging.list.entries.find((renditionEntry: RenditionEntry) => renditionEntry.entry.id.toLowerCase() === renditionId);
+    isCustomViewerExtension(extension: string): boolean {
+        const extensions = this.externalExtensions || [];
 
-        return rendition?.entry?.status?.toString() === 'CREATED' || false;
+        if (extension && extensions.length > 0) {
+            extension = extension.toLowerCase();
+            return extensions.flat().indexOf(extension) >= 0;
+        }
+
+        return false;
     }
+
+
 }
