+++ lib/core/src/lib/viewer/viewer.module.ts
@@ -43,6 +43,7 @@ import { ViewerExtensionDirective } from './directives/viewer-extension.directiv
 import { ViewerToolbarActionsComponent } from './components/viewer-toolbar-actions.component';
 import { DirectiveModule } from '../directives/directive.module';
 import { A11yModule } from '@angular/cdk/a11y';
+import { AlfrescoViewerComponent } from "./components/alfresco-viewer.component";
 
 @NgModule({
     imports: [
@@ -61,6 +62,7 @@ import { A11yModule } from '@angular/cdk/a11y';
     declarations: [
         PdfPasswordDialogComponent,
         ViewerComponent,
+        AlfrescoViewerComponent,
         ImgViewerComponent,
         TxtViewerComponent,
         MediaPlayerComponent,
@@ -77,6 +79,7 @@ import { A11yModule } from '@angular/cdk/a11y';
     ],
     exports: [
         ViewerComponent,
+        AlfrescoViewerComponent,
         ImgViewerComponent,
         TxtViewerComponent,
         MediaPlayerComponent,
